requirements.txt
flask
google-genai
PyPDF2